Start moving things to better places.
