+++ src/dev_utilities/mod.rs
@@ -1,0 +1,2 @@
+//! Utilities to be used when developing backends and middleware.
+pub mod vecstorage;
