+++ src/middleware/zero_init.rs
@@ -4,7 +4,7 @@
 //! before calling the plugin and you are using the `Polyphony` middleware.
 
 use asprim::AsPrim;
-use backend::{Plugin, Transparent};
+use crate::{Plugin, Transparent};
 use num_traits::Float;
 
 /// Set all output values to 0 before calling `render_buffer` on the "child".
